Remove mina-graphql-client sub-crate, use mina-sdk-rust instead
The mina-graphql-client sub-crate duplicated code already present in
internal-log-fetcher (authentication, graphql queries, schema). This
replaces the sub-crate with:

- mina-sdk (new shared Mina Rust SDK) for standard daemon queries
- ITN-specific logic (auth, internal logs, post_graphql) inlined into
  MinaServer directly

All 6 existing tests pass with original precomputed signature values.
