+++ internal-log-fetcher/Cargo.toml
@@ -30,7 +30,9 @@ tokio = { version = "1", features = ["full"] }
 tracing = "0.1"
 tracing-subscriber = "0.3"
 warp = "0.3"
-mina-graphql-client = { path = "./mina-graphql-client" }
+# TODO: switch to git dependency once mina-sdk-rust repo is published
+# mina-sdk = { git = "https://github.com/MinaProtocol/mina-sdk-rust.git", branch = "master" }
+mina-sdk = { path = "/home/darek/work/minaprotocol/mina-sdk-rust" }
 
 [dev-dependencies]
 httpmock = "0.8.0-alpha.1"
