--- internal-log-fetcher/mina-graphql-client/Cargo.toml
@@ -1,37 +1,0 @@
-[package]
-name = "mina-graphql-client"
-version = "0.1.0"
-authors = ["Bruno Deferrari <bruno.deferrari@viablesystems.io>, o1labs <o1labs.org>"]
-edition = "2021"
-license = "Apache-2.0"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-anyhow = "1.0"
-base64 = "0.21"
-chrono = "0.4"
-ed25519-dalek = "1"
-futures-util = "0.3"
-graphql_client = { version = "0.12" }
-object_store = { version = "0.6", features = [
-  "reqwest",
-  "gcp",
-  "serde",
-  "serde_json",
-  "aws",
-] }
-rand = "0.7"
-reqwest = { version = "0.11", features = ["json"] }
-serde = "1.0"
-serde_json = { version = "1", features = ["raw_value"] }
-structopt = "0.3"
-tokio = { version = "1", features = ["full"] }
-tracing = "0.1"
-tracing-subscriber = "0.3"
-warp = "0.3"
-url = "2.5.4"
-
-
-[dev-dependencies]
-httpmock = "0.8.0-alpha.1"
