--- internal-log-fetcher/mina-graphql-client/graphql/auth_query.graphql
@@ -1,7 +1,0 @@
-query AuthQuery {
-    auth {
-        serverUuid
-        signerSequenceNumber,
-        isBlockProducer
-    }
-}
