--- internal-log-fetcher/mina-graphql-client/graphql/flush_internal_logs_mutation.graphql
@@ -1,3 +1,0 @@
-mutation FlushInternalLogsQuery($logId: Int!) {
-    flushInternalLogs(endLogId: $logId)
-}
