--- internal-log-fetcher/mina-graphql-client/graphql/internal_logs_query.graphql
@@ -1,12 +1,0 @@
-query InternalLogsQuery($logId: Int!) {
-    internalLogs(startLogId: $logId) {
-        id
-        timestamp
-        message
-        metadata {
-            item
-            value
-        }
-        process
-    }
-}
