--- internal-log-fetcher/mina-graphql-client/graphql/reset_zkapp_soft_limit.graphql
@@ -1,3 +1,0 @@
-mutation ResetZkappSoftLimitQuery {
-    zkAppCommandLimit(limit: Null)
-}
