--- internal-log-fetcher/mina-graphql-client/graphql/schedule_zkapp_commands.graphql
@@ -1,3 +1,0 @@
-mutation ScheduleZkappCommandsQuery ($input: ZkappCommandsDetails!) {
-    scheduleZkappCommands(input: $input)
-}
