--- internal-log-fetcher/mina-graphql-client/graphql/schema.graphql
@@ -1,210 +1,0 @@
-schema {
-  query: query
-  mutation: mutation
-}
-
-"""uint64 encoded as a json string representing an ammount of currency"""
-scalar CurrencyAmount
-
-"""uint64 encoded as a json string representing a fee"""
-scalar Fee
-
-"""Update to gating config and added peers"""
-input GatingUpdate {
-  """Peers to connect to"""
-  addedPeers: [NetworkPeer!]!
-
-  """If true, resets added peers to an empty list (including seeds)"""
-  cleanAddedPeers: Boolean!
-
-  """
-  If true, no connections will be allowed unless they are from a trusted peer
-  """
-  isolate: Boolean!
-
-  """
-  Peers we will never allow connections from (unless they are also trusted!)
-  """
-  bannedPeers: [NetworkPeer!]!
-
-  """Peers we will always allow connections from"""
-  trustedPeers: [NetworkPeer!]!
-}
-
-type ItnAuth {
-  """Uuid of the ITN GraphQL server"""
-  serverUuid: String!
-
-  """Sequence number for the signer of the auth query"""
-  signerSequenceNumber: UInt16!
-
-  """Libp2p port"""
-  libp2pPort: UInt16!
-
-  """Peer id"""
-  peerId: String!
-
-  """Is the node a block producer"""
-  isBlockProducer: Boolean!
-}
-
-type ItnLog {
-  """the log ID"""
-  id: Int!
-
-  """timestamp of the log"""
-  timestamp: String!
-
-  """the log message"""
-  message: String!
-
-  """metadata for the log"""
-  metadata: [logMetadatum!]!
-
-  """optional id of the process"""
-  process: String
-}
-
-"""Arbitrary JSON"""
-scalar Json
-
-type logMetadatum {
-  """metadatum item"""
-  item: String!
-
-  """metadatum value"""
-  value: Json!
-}
-
-
-"""Base58Check-encoded private key"""
-scalar PrivateKey
-
-"""Public key in Base58Check format"""
-scalar PublicKey
-
-type mutation {
-  schedulePayments(
-    """Payments details"""
-    input: PaymentsDetails!
-  ): String!
-
-  scheduleZkappCommands (
-    """Payments details"""
-    input: ZkappCommandsDetails!
-  ): String!
-  stopPayments(
-    """Payment scheduler handle"""
-    handle: String!
-  ): String!
-  updateGating(
-    """Gating update"""
-    input: GatingUpdate!
-  ): String!
-
-  """Returns number of logs deleted from queue"""
-  flushInternalLogs(
-    """Greatest log ID to be deleted"""
-    endLogId: Int!
-  ): String!
-
-  """ Sets zkapp commands soft limit """
-  zkAppCommandLimit(
-    limit: [Int]
-  ): String!
-}
-
-"""Network identifiers for another protocol participant"""
-input NetworkPeer {
-  libp2pPort: Int!
-
-  """IP address of the remote host"""
-  host: String!
-
-  """base58-encoded peer ID"""
-  peerId: String!
-}
-
-"""Keys and other information for scheduling payments"""
-input PaymentsDetails {
-  """Length of scheduler run"""
-  durationInMinutes: Int!
-
-  """Frequency of transactions"""
-  transactionsPerSecond: Float!
-
-  """Memo, up to 32 characters"""
-  memo: String!
-
-  """Maximum fee"""
-  feeMax: Fee!
-
-  """Minimum fee"""
-  feeMin: Fee!
-
-  """Amount for payments"""
-  amount: CurrencyAmount!
-
-  """Public key of receiver of payments"""
-  receiver: PublicKey!
-
-  """Private keys of accounts to send from"""
-  senders: [PrivateKey!]!
-}
-
-input ZkappCommandsDetails {
-
-  maxAccountUpdates: Int!
-  
-  maxCost: Boolean!
-
-  accountQueueSize: Int!
-
-  deploymentFee: Int!
-
-  maxFee: Int!
-
-  minFee: Int!
-
-  initBalance: Int!
-
-  maxNewZkappBalance: Int!
-
-  minNewZkappBalance: Int!
-
-  maxBalanceChange: Int!
-
-  minBalanceChange: Int!
-
-  noPrecondition: Boolean!
-
-  memoPrefix: String!
-
-  durationMin: Int!
-
-  tps: Float!
-
-  numNewAccounts: Int!
-
-  numZkappsToDeploy: Int!
-
-  """Private keys of accounts to send from"""
-  feePayers: [String!]!
-}
-
-type query {
-  """Uuid for GraphQL server, sequence number for signing public key"""
-  auth: ItnAuth!
-
-  """Slots won by a block producer for current epoch"""
-  slotsWon: [Int!]!
-
-  """Internal logs generated by the daemon"""
-  internalLogs(
-    """Least log ID to start with"""
-    startLogId: Int!
-  ): [ItnLog!]!
-}
-
-"""String representing a uint16 number in base 10"""
-scalar UInt16
