--- internal-log-fetcher/mina-graphql-client/src/lib.rs
@@ -1,7 +1,0 @@
-mod authentication;
-mod graphql;
-mod mina_client;
-
-pub use graphql::internal_logs_query::InternalLogsQueryInternalLogs;
-pub use graphql::schedule_zkapp_commands_query::ZkappCommandsDetails;
-pub use mina_client::{MinaClientConfig, MinaGraphQLClient};
