+++ internal-log-fetcher/src/log_entry.rs
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::utils::convert_timestamp_to_float;
-use mina_graphql_client::InternalLogsQueryInternalLogs;
+use crate::graphql::internal_logs_query::InternalLogsQueryInternalLogs;
 use serde::ser::{Serialize, SerializeSeq, Serializer};
 use serde_json::Map;
 use std::error::Error;
